Add support for reader writer structs
This is a big change in zig-avro. We can now generate structs from given `.avsc`
files, allowing more ease of use and reducing error-proneness when typing out
the structs. There are still some rough edges, and some known caveats for the
time being. Namely:

- We hardcode folders for input schemas and output code
- We generate all structs given from all `.avsc` files into one file
- We still don't support unions perfectly

Ok, so for the changes. We introduce some new helper types and functions:

**`avro.Iterator`**:

This is the magic sauce that allows us to have transparent reading and writing
of an iterable. What happens is that this type introduces an interface that
_can_ be implemented userspace to allow for custom write logic. This interface
consists of two functions, `next` and `iterator`, where the former is the
implementation of the interface, and the latter is the initialization. If both
exist, `zig-avro` will use these in the `avro.Writer` functions and still allow
us the zero copy flow of data.

**`avro.Map`**:

This is an abstraction over our existing `avro.Reader.Map` and the
new `avro.Iterator`, which see.

**`avro.Array`**:

This is an abstraction over our existing `avro.Reader.Array` and the
new `avro.Iterator`, which see.