Expand support for union generation
This was a little tough. We need to make sure we handle a `union{null, Foo}` as a
?Foo, and also kind of handle nullability for unions with more variants, like
`union{null, Foo, Bar, Baz}`, where avro defines `null` as a decent value for
the latter. This doesn't really work in Zig, as we need a useful void
variant. As such we do some checking for `null` as a variant in the union, as
that by convention should be determined as the null value. In that case we
default to the `.null` variant, as this encapsulates `void`.

This should work just fine, and there is a test showcasing a rather complex file
using many of these shenanigans.