Fix array encoding bug and add tests
This bug looks like it has been here for a while.  When encoding an
array with a known length of 0, it adds a block of length 0 before
ending the array with a block of length zero to signal the end of
the array.  Since the first 0 indicates the end of the array, the
second 0 is corrupt data.